perf(mem): use float32 for full-image buffers in hdr/drizzle/analysis
These paths allocated full-image-sized float64 arrays where float32 is
sufficient, doubling peak memory on large frames:

- hdr._weighted_average: float64 accumulator + per-frame .astype(float64).
  Exposure fusion blends only a handful of frames; float32 round-off ~1e-7.
- drizzle CPU fallback: float64 homogeneous coords (~1.75GB at 73MP) and the
  scaled output/weight accumulators. Pixel coords < 1e4 resolve to ~1e-3 px
  in float32; accumulation is over dozens of frames, not millions.
- background gaussian fallback: float64 temp of the full background model.
- fwhm_map / aperture_photometry / tilt_analysis: grayscale was upcast to
  float64 (np.mean of float32 is already float32); star-detection math
  (median/MAD/center_of_mass) is unaffected by float32.

Also removed a dead 'n' assignment and wrapped two pre-existing long lines.
54 hdr/drizzle/background/analysis tests pass.

+++ astraios/core/analysis/aperture_photometry.py
@@ -301,9 +301,9 @@ def run_photometry(
         params = PhotometryParams()
 
     if image.ndim == 3:
-        gray = np.mean(image, axis=0).astype(np.float64)
+        gray = np.mean(image, axis=0).astype(np.float32)
     else:
-        gray = image.astype(np.float64)
+        gray = image.astype(np.float32)
 
     # Try photutils first (most feature-complete)
     try:
